requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
PyJWT==2.4.0
cryptography==36.0.0
requests==2.26.0
